+++ tests/trainer/test_trainer_tricks.py
@@ -15,6 +15,7 @@ import torch
 from torch.utils.data import DataLoader, RandomSampler, SequentialSampler
 
 from pytorch_lightning import Trainer
+from pytorch_lightning.trainer.states import RunningStage
 from tests.base import EvalModelTemplate
 
 
@@ -107,7 +108,7 @@ def test_overfit_batch_limits(tmpdir):
     # ------------------------------------------------------
     # run tests for both val and test
     # ------------------------------------------------------
-    for split in ["val", "test"]:
+    for split in (RunningStage.VALIDATING, RunningStage.TESTING):
 
         # ------------------------------------------------------
         # test overfit_batches as percent
@@ -134,7 +135,7 @@ def test_overfit_batch_limits(tmpdir):
         # ------------------------------------------------------
         # test limit_xxx_batches as percent AND int
         # ------------------------------------------------------
-        if split == "val":
+        if split == RunningStage.VALIDATING:
             loader_num_batches, dataloaders = Trainer(limit_val_batches=0.1)._reset_eval_dataloader(split, model=model)
             assert loader_num_batches[0] == int(0.1 * len(val_loader))